+++ ttmp/2026/04/25/PYXIS-CSSVD-JS-LIB--implement-pyxis-css-visual-diff-javascript-userland-library/scripts/lib/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  slug: require('./slug.js'),
+  storybook: require('./storybook.js'),
+  artifacts: require('./artifacts.js'),
+  registry: require('./registry.js'),
+}