+++ ttmp/2026/04/25/PYXIS-CSSVD-JS-LIB--implement-pyxis-css-visual-diff-javascript-userland-library/scripts/lib/storybook.js
@@ -1,0 +1,10 @@
+function trimTrailingSlash(value) {
+  return String(value || '').replace(/\/+$/, '')
+}
+
+function storybookIframeUrl(baseUrl, storyId) {
+  var base = trimTrailingSlash(baseUrl)
+  return base + '/iframe.html?id=' + encodeURIComponent(storyId) + '&viewMode=story'
+}
+
+module.exports = { storybookIframeUrl }